fix: harden route transforms and dev lifecycle
