+++ rslib.config.ts
@@ -5,16 +5,30 @@ import {
 } from '@rsbuild/config/rslib.config.js';
 import { defineConfig } from '@rslib/core';
 const config = defineConfig({
-  source: {
-    entry: {
-      index: './src/index.ts',
-      'parallel-route-transform-worker':
-        './src/parallel-route-transform-worker.ts',
-      'templates/entry.server': './src/templates/entry.server.tsx',
-      'templates/entry.client': './src/templates/entry.client.tsx',
+  lib: [
+    {
+      ...esmConfig,
+      source: {
+        entry: {
+          index: './src/index.ts',
+          'parallel-route-transform-worker':
+            './src/parallel-route-transform-worker.ts',
+          'templates/entry.server': './src/templates/entry.server.tsx',
+          'templates/entry.client': './src/templates/entry.client.tsx',
+        },
+      },
     },
-  },
-  lib: [esmConfig, cjsConfig],
+    {
+      ...cjsConfig,
+      source: {
+        entry: {
+          index: './src/index.ts',
+          'templates/entry.server': './src/templates/entry.server.tsx',
+          'templates/entry.client': './src/templates/entry.client.tsx',
+        },
+      },
+    },
+  ],
   tools: {
     rspack: {
       externals: [
